[HER-104] Change score type in PDetection

+++ src/Business/Detection/Fetching/Models/PDetection.cs
@@ -7,7 +7,7 @@ namespace Business.Detection.Fetching.Models
     {
         public int? Id { get; set; }
 
-        public int? Score { get; set; }
+        public float? Score { get; set; }
 
         public string Class { get; set; }
